layers:Disable memory access callback
This is a conservative measure to simplify landing and integration of
the memory access checks. For now turning off the callback and also
disabling some draw-time checks that seem to have ~10-20% perf hit in
limited testing. Turning off callback also means turning off tests that
rely on the callback.
Will continue advancing this code and testing with checks enabled to
reach desired functionality before enabling code on master.